% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateColorPlot.R
\name{aggregateColorPlot}
\alias{aggregateColorPlot}
\title{Plot aggregate soil color data}
\usage{
aggregateColorPlot(
  x,
  print.label = TRUE,
  label.font = 1,
  label.cex = 0.65,
  label.orientation = c("v", "h"),
  buffer.pct = 0.02,
  print.n.hz = FALSE,
  rect.border = "black",
  horizontal.borders = FALSE,
  horizontal.border.lwd = 2,
  x.axis = TRUE,
  y.axis = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{list}, results from \code{aqp::aggregateColor()}}

\item{print.label}{logical, print Munsell color labels inside of rectangles, only if they fit}

\item{label.font}{font specification for color labels}

\item{label.cex}{font size for color labels}

\item{label.orientation}{label orientation, \code{v} for vertical or \code{h} for horizontal}

\item{buffer.pct}{extra space between labels and color rectangles}

\item{print.n.hz}{optionally print the number of horizons below Munsell color labels}

\item{rect.border}{color for rectangle border}

\item{horizontal.borders}{optionally add horizontal borders between bands of color}

\item{horizontal.border.lwd}{line width for horizontal borders}

\item{x.axis}{logical, add a scale and label to x-axis?}

\item{y.axis}{logical, add group labels to y-axis?}

\item{...}{additional arguments passed to \code{plot}}
}
\value{
nothing, function called for graphical output
}
\description{
Generate a plot from summaries generated by \code{aqp::aggregateColor()}.
}
\details{
Tutorial at \url{http://ncss-tech.github.io/AQP/sharpshootR/aggregate-soil-color.html}.
}
\author{
D.E. Beaudette
}
\keyword{hplots}
