\name{amador}
\alias{amador}
\docType{data}
\title{SSURGO Data Associated with the Amador Soil Series}
\description{SSURGO Data Associated with the Amador Soil Series}
\usage{data(amador)}
\format{
  A subset of data taken from the "component" table of SSURGO
  \describe{
    \item{\code{mukey}}{map unit key}
    \item{\code{compname}}{component name}
    \item{\code{comppct_r}}{component percentage}
  }
}
\source{USDA-NRCS SSURGO Database}
\keyword{datasets}
