% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAF.R
\name{plotDAF}
\alias{plotDAF}
\alias{DAF.C}
\alias{DAF.G}
\alias{DAF.D}
\alias{DAF.B}
\title{Demonstrate Declining Availability Functions}
\usage{
plotDAF(s = c("B", "C", "D", "G"), res = 0.01, ...)

DAF.C(x)

DAF.G(x)

DAF.D(x)

DAF.B(x)
}
\arguments{
\item{s}{character, selection of declining availability function by name ('B', 'C', 'D', 'G')}

\item{res}{numeric, curve resolution}

\item{...}{additional arguments to \code{plot()}}

\item{x}{ordered sequence of VWC in (0, 1).}
}
\value{
AET/PET ratio in (0, 10)

AET/PET ratio in (0, 10)

AET/PET ratio in (0, 10)

AET/PET ratio in (0, 10)
}
\description{
This function ...
}
\keyword{internal}
