% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyWB.R
\name{dailyWB}
\alias{dailyWB}
\title{Simple Daily Water Balance}
\usage{
dailyWB(x, daily.data, id, S_0 = 0.5)
}
\arguments{
\item{x}{\code{data.frame}, required columns include:
\itemize{
\item \code{sat}
\item \code{fc}
\item \code{pwp}
\item \code{thickness}
\item \code{a.ss}
\item "id"
}}

\item{daily.data}{\code{data.frame}, required columns include:
\itemize{
\item \code{date}
\item \code{PPT}
\item \code{PET}
}}

\item{id}{character, name of column in \code{x} that is used to identify records}

\item{S_0}{fraction of water storage filled at time = 0 (range: 0-1)}
}
\value{
a \code{data.frame}
}
\description{
A simple daily water balance.
}
