% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dueling.dendrograms.R
\name{dueling.dendrograms}
\alias{dueling.dendrograms}
\title{Dueling Dendrograms}
\usage{
dueling.dendrograms(
  p.1,
  p.2,
  lab.1 = "D1",
  lab.2 = "D2",
  cex.nodelabels = 0.75,
  arrow.length = 0.05
)
}
\arguments{
\item{p.1}{left-hand phylo-class dendrogram}

\item{p.2}{right-hand phylo-class dendrogram}

\item{lab.1}{left-hand title}

\item{lab.2}{right-hand title}

\item{cex.nodelabels}{character expansion size for node labels}

\item{arrow.length}{arrow head size}
}
\value{
nothing is returned, function is called to generate graphical output
}
\description{
Graphically compare two related dendrograms
}
\details{
Connector arrows are used to link nodes from the left-hand dendrogram to the right-hand dendrogram.
}
\author{
D.E. Beaudette
}
\keyword{hplots}
