% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moran_I_ByRaster.R
\name{ESS_by_Moran_I}
\alias{ESS_by_Moran_I}
\title{Estimate Effective Sample Size}
\usage{
ESS_by_Moran_I(n, rho)
}
\arguments{
\item{n}{sample size}

\item{rho}{Global Moran's I}
}
\value{
numeric; estimated Effective Sample Size
}
\description{
Estimation of effective sample size (ESS). See Fortin & Dale 2005, p. 223, Equation 5.15 using global Moran's I as 'rho'.
}
\references{
Fortin, M.J. and Dale, M.R.T. (2005) Spatial Analysis: A Guide for Ecologists. Cambridge University Press, Cambridge, 1-30.
}
\author{
D.E. Beaudette
}
