% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HenryTimeLine.R
\name{HenryTimeLine}
\alias{HenryTimeLine}
\title{Sensor Data Timeline from Henry Mount Soil and Water DB}
\usage{
HenryTimeLine(sensor_data, ...)
}
\arguments{
\item{sensor_data}{`soiltemp`, `soilVWC`, or related data returned by \code{soilDB::fetchHenry()}}

\item{...}{additional arguments to \code{latticeExtra::segplot}}
}
\value{
a \code{lattice} graphics object
}
\description{
This function generates a simple chart of start/end dates for a set of sensor data returned by \code{soilDB::fetchHenry}.
}
\note{
This function does not symbolize sections of missing data between the first and last record.
}
\author{
D.E. Beaudette
}
