% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMineralSoilMaterial.R
\name{isMineralSoilMaterial}
\alias{isMineralSoilMaterial}
\title{Mineral Soil Material Criteria from 12th Ed. of KST}
\usage{
isMineralSoilMaterial(soc, clay, saturation = TRUE)
}
\arguments{
\item{soc}{soil organic carbon percent by mass}

\item{clay}{clay content percent by mass}

\item{saturation}{logical, cumulative saturation 30+ days}
}
\value{
`data.frame` of criteria test results
}
\description{
Evaluate mineral soil material criteria based on soil organic carbon, clay content, and length of saturation.
}
