% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinAdjacency.R
\name{joinAdjacency}
\alias{joinAdjacency}
\title{Join Document Adjacency}
\usage{
joinAdjacency(x, vars = c("l_musym", "r_musym"))
}
\arguments{
\item{x}{\code{data.frame} or similar object, each row represents a single shared edge (typically \code{sf} LINESTRING feature)}

\item{vars}{a vector of two characters naming columns containing "left", and "right" map unit symbols}
}
\value{
A weighted adjacency matrix is returned, suitable for plotting directly with \code{plotSoilRelationGraph()}.
}
\description{
Convert a set of line segment "join decisions" into a weighted adjacency matrix describing which map unit symbols touch.
}
\seealso{
\code{\link[=plotSoilRelationGraph]{plotSoilRelationGraph()}}
}
\author{
D.E. Beaudette
}
