% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moistureStateProportions.R
\name{moistureStateProportions}
\alias{moistureStateProportions}
\title{Compute moisture state proportions}
\usage{
moistureStateProportions(x, id = "series", step = c("month", "week", "doy"))
}
\arguments{
\item{x}{\code{data.frame} created by \code{dailyWB_SSURGO()}}

\item{id}{character, column name identifying sites, soils, or soil series}

\item{step}{time step, one of 'month', 'week', or 'doy'}
}
\value{
\code{data.frame}
}
\description{
Compute moisture state proportions
}
