% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moistureStateProportions.R
\name{moistureStateStats}
\alias{moistureStateStats}
\title{Statistics on Soil Moisture State}
\usage{
moistureStateStats(x, id = "compname")
}
\arguments{
\item{x}{\code{data.frame}, created by \code{\link[=moistureStateProportions]{moistureStateProportions()}}}

\item{id}{name of ID column}
}
\value{
\code{data.frame} containing the most-likely moisture state and Shannon entropy.
}
\description{
Statistics on Soil Moisture State
}
