% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moistureStateThreshold.R
\name{moistureStateThreshold}
\alias{moistureStateThreshold}
\title{Apply a threshold to soil moisture states}
\usage{
moistureStateThreshold(
  x,
  id = "series",
  threshold = "moist",
  operator = c("<", ">", "==", "<=", ">=")
)
}
\arguments{
\item{x}{\code{data.frame} created by \code{\link[=dailyWB_SSURGO]{dailyWB_SSURGO()}}}

\item{id}{character, column name identifying sites, soils, or soil series}

\item{threshold}{moisture state threshold, see \code{\link{estimateSoilMoistureState}}}

\item{operator}{one of "<", ">", "==", "<=", or ">="}
}
\value{
\code{data.frame}
}
\description{
Apply a threshold to soil moisture states
}
\author{
D.E. Beaudette
}
