% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{OSDexamples}
\alias{OSDexamples}
\title{Example output from soilDB::fetchOSD()}
\format{
An object of class \code{list} of length 16.
}
\usage{
data(OSDexamples)
}
\description{
These example data are used to test various functions in this package when network access may be limited.
}
\keyword{datasets}
