% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSS2LL.R
\name{PLSS2LL}
\alias{PLSS2LL}
\alias{PLSS2LL_1,PLSS2LL_oneline}
\title{PLSS2LL}
\usage{
PLSS2LL(p)
}
\arguments{
\item{p}{dataframe with chunks of PLSS coordinates}
}
\value{
A \code{data.frame} of PLSS codes and coordinates.
}
\description{
Fetch lattitude and longitude centroid coordinates for coded PLSS information from the BLM PLSS web service.
}
\note{
This function expects that the dataframe will have a 'plssid' column generated by the \code{formatPLSS} function. Requires the following packages: \code{httr}, and \code{jsonlite}.
}
\seealso{
\code{\link{PLSS2LL_oneline}},\code{\link{LL2PLSS}}, \code{\link{formatPLSS}}
}
\author{
D.E. Beaudette, Jay Skovlin
}
