% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_SSURGO_hydro_data.R
\name{prepare_SSURGO_hydro_data}
\alias{prepare_SSURGO_hydro_data}
\title{Get and prepare basic soil hydraulic parameters from SSURGO via SDA}
\usage{
prepare_SSURGO_hydro_data(cokeys, max.depth)
}
\arguments{
\item{cokeys}{vector of component keys (cokey) in current SSURGO snapshot}

\item{max.depth}{target depth of aggregation, corrected later by real soil depth as reported by \code{slab()}}
}
\value{
a \code{list} containing:
\itemize{
\item \code{SPC}: \code{SoilProfileCollection}
\item \code{agg}: aggregate representation of hydraulic parameters
}
}
\description{
Get and prepare basic soil hydraulic parameters from SSURGO via SDA
}
\author{
D.E. Beaudette
}
