% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcileOSDGeomorph.R
\name{reconcileOSDGeomorph}
\alias{reconcileOSDGeomorph}
\title{Reconcile IDs between a SPC and associated geomorphic proportion table}
\usage{
reconcileOSDGeomorph(
  x,
  selection = c("hillpos", "geomcomp", "flats", "mtnpos", "terrace", "shape_across",
    "shape_down")
)
}
\arguments{
\item{x}{resulting list from \code{soilDB::fetchOSD(..., extended = TRUE)}}

\item{selection}{character, name of geomorphic proportion table}
}
\value{
a \code{list} with subset \code{SoilProfileCollection} and \code{data.frame} of geomorphic proportions, \code{selection} is preserved as an attribute.
}
\description{
This function can assist with linked visualizations that include soil morphology data stored in a \code{SoilProfileCollection} and geomorphic proportions stored in a \code{data.frame}, as returned by \code{soilDB::fetchOSD()}.
}
\author{
D.E. Beaudette
}
