% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.by.poly.R
\name{sample.by.poly}
\alias{sample.by.poly}
\title{Sample a Single Polygon at a Target Density}
\usage{
sample.by.poly(
  p,
  n.pts.per.ac = 1,
  min.samples = 5,
  sampling.type = "regular",
  p4s = NULL
)
}
\arguments{
\item{p}{\code{SpatVector} containing a single polygon feature}

\item{n.pts.per.ac}{numeric, sampling density in "points per acre"}

\item{min.samples}{integer, enforced minimum number of samples per polygon}

\item{sampling.type}{sampling.type character, sampling type passed to \code{\link[terra:sample]{terra::spatSample()}}}

\item{p4s}{output from \code{\link[terra:crs]{terra::crs()}}}
}
\value{
\code{SpatVector} of sample points
}
\description{
Sample a single \code{SpatVector} polygon at a target density in "samples / acre".
}
\author{
D.E. Beaudette
}
\keyword{manip}
