\name{sharpshootR-package}
\alias{sharpshootR-package}
\alias{sharpshootR}
\docType{package}
\title{A collection of functions to support soil survey}
\description{This package contains mish-mash of functionality and sample data related to the daily business of soil survey operations with the USDA-NRCS. Many of the functions are highly specialized and inherit default arguments from the names used by the various NCSS (National Cooperative Soil Survey) databases. A detailed description of this package with links to associated tutorials can be found at the \href{https://github.com/ncss-tech/sharpshootR}{project website}.}
