% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleWB.R
\name{simpleWB}
\alias{simpleWB}
\title{Title}
\usage{
simpleWB(PPT, PET, D, thickness, sat, fc, S_0 = 0.5, a.ss = 0.03)
}
\arguments{
\item{PPT}{precipitation series (mm)}

\item{PET}{potential ET series (mm)}

\item{D}{dates (???)}

\item{thickness}{soil thickness (cm)}

\item{sat}{volumeric water content at saturation (satiated water content)}

\item{fc}{volumetric water content at field capacity (typically 1/3 bar suction)}

\item{...}{}
}
\value{

}
\description{
Title
}
\details{
If \code{awc} has not been adjusted for coarse fragments, then the adjustment can be made to \code{thickness}.
}
