% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizHillslopePosition.R
\name{vizHillslopePosition}
\alias{vizHillslopePosition}
\title{Visual Summary of Hillslope Position}
\usage{
vizHillslopePosition(
  x,
  s = NULL,
  annotations = TRUE,
  annotation.cex = 0.75,
  clust = TRUE,
  dend.size = 5,
  dend.type = c("rectangle", "triangle"),
  cols = c("#5d74a5", "#b0cbe7", "#fef7c7", "#eba07e", "#a8554e"),
  ...
)
}
\arguments{
\item{x}{\code{data.frame} as created by \code{soilDB::fetchOSD(..., extended = TRUE)}}

\item{s}{an optional soil series name, highlighted in the figure}

\item{annotations}{logical, add number of record and normalized Shannon entropy values}

\item{annotation.cex}{annotation label scaling factor}

\item{clust}{logical, order rows using divisive hierarchical clustering and include dendrogram?}

\item{dend.size}{numeric, space reserved for dendrogram when \code{clust = TRUE}: values between 3-10 are about right}

\item{dend.type}{character, one of 'rectangle' or 'triangle', when \code{clust = TRUE}}

\item{cols}{vector of colors}

\item{\dots}{additional arguments to \code{\link[=iterateHydOrder]{iterateHydOrder()}}: \verb{target = 0.9, maxIter = 20, j.amount = 0.001, verbose = FALSE}}
}
\value{
A \code{list} with the following elements:
\itemize{
\item \code{fig}: lattice object (the figure)
\item \code{order}: 1D ordering from \code{cluster::diana}
\item \code{clust}: \code{hclust} object
\item \code{match.rate}: fraction of series matching target hydrologic ordering, after clustering + rotation
}
}
\description{
A unique display of hillslope position probability.
}
\details{
See the \href{http://ncss-tech.github.io/AQP/soilDB/soil-series-query-functions.html}{Soil Series Query Functions} tutorial for more information.
}
\note{
Default colors are from \code{PNWColors::pnw_palette('Shuksan2')}. Original colors: \code{c("#2B83BA", "#ABDDA4", "#FFFFBF", "#FDAE61", "#D7191C")}.
}
\author{
D.E. Beaudette
}
