% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyWB.R
\name{monthlyWB}
\alias{monthlyWB}
\title{Monthly Water Balances}
\usage{
monthlyWB(
  AWC,
  PPT,
  PET,
  S_init = AWC,
  starting_month = 1,
  rep = 1,
  keep_last = FALSE
)
}
\arguments{
\item{AWC}{available water-holding capacity (mm), typically thickness (mm) * awc (fraction)}

\item{PPT}{time-series of monthly PPT (mm), calendar year ordering}

\item{PET}{time-series of monthly PET (mm), calendar year ordering}

\item{S_init}{initial fraction of \code{AWC} filled with water}

\item{starting_month}{starting month index, 1=January, 9=September}

\item{rep}{number of cycles to run water balance}

\item{keep_last}{keep only the last iteration of the water balance}
}
\value{
a \code{data.frame} with the following elements:

\itemize{
\item{PPT: }{monthly PPT (mm)}
\item{PET: }{monthly PET (mm)}
\item{U: }{monthly surplus (mm)}
\item{S: }{monthly soil moisture storage (mm)}
\item{ET: }{monthly AET (mm)}
\item{D: }{monthly deficit (mm)}
\item{month: }{month number}
\item{mo: }{month label}
}
}
\description{
Perform a monthly water balance by "leaky bucket" model, inspired by code from \code{bucket.sim} of \code{hydromad} package, as defined in Bai et al., (2009) (model "SMA_S1"). The plant available water-holding storage (soil thickness * awc) is used as the "bucket capacity". All water in excess of this capacity is lumped into a single "surplus" term.
}
\details{
A number of important assumptions are made by this approach:
\itemize{
\item the concept of field capacity is built into the specified bucket size
\item the influence of aquitards or local terrain cannot be integrated into this model
\item interception is not used in this model
}
}
\examples{

# 4" water storage ~ 100mm

# AWC in mm
AWC <- 100

# monthly PET and PPT in mm
PET <- c(0,0,5,80,90,120,130,140,110,90,20,5)
PPT <- c(0, 150, 200, 120, 20, 0, 0, 0, 10, 20, 30, 60)

# run water balance
# start with soil AWC "empty"
(x.wb <- monthlyWB(AWC, PPT, PET, S_init = 0))

# plot the results
op <- par(no.readonly = TRUE)

par(mar=c(4,4,2,1), bg = 'white')
plotWB(WB = x.wb)

# compute fraction of AWC filled after the last month of simulation
(last.S <- x.wb$S[12] / AWC)

# re-run the water balance with this value
(x.wb <- monthlyWB(AWC, PPT, PET, S_init = last.S))

# interesting...
par(mar=c(4,4,2,1), bg = 'white')
plotWB(WB = x.wb)

# note: consider using `rep = 3, keep_last = TRUE` 
# to "warm-up" the water balance first

par(op)


}
\references{
Arkley R, Ulrich R. 1962. The use of calculated actual and potential evapotranspiration for estimating potential plant growth. Hilgardia 32(10):443-469.

Bai, Y., T. Wagener, P. Reed (2009). A top-down framework for watershed model evaluation and selection under uncertainty. Environmental Modelling and Software 24(8), pp. 901-916.

Farmer, D., M. Sivapalan, Farmer, D. (2003). Climate, soil and vegetation controls upon the variability of water balance in temperate and semiarid landscapes: downward approach to water balance analysis. Water Resources Research 39(2), p 1035.
}
