% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeomorphCrossSection.R
\name{plotGeomorphCrossSection}
\alias{plotGeomorphCrossSection}
\title{Present a \code{SoilProfileCollection} aligned to a geomorphic summary as cross-section.}
\usage{
plotGeomorphCrossSection(
  x,
  type = c("line", "bar"),
  g = "hillpos",
  clust = TRUE,
  col = c("#377EB8", "#4DAF4A", "#984EA3", "#FF7F00", "#E41A1C"),
  ...
)
}
\arguments{
\item{x}{resulting list from \code{soilDB::fetchOSD(..., extended = TRUE)}}

\item{type}{character, 'line' for line plot or 'bar' for barplot of geomorphic proportions}

\item{g}{character, select a geomorphic summary. Currently 'hillpos' (2D hillslope position) is the only supported choice.}

\item{clust}{logical, use clustering order of geomorphic proportions (\code{TRUE}) or exact hydrologic ordering (\code{FALSE}), see \code{\link[=hydOrder]{hydOrder()}}}

\item{col}{character vector of colors}

\item{\dots}{additional arguments to \code{\link[=iterateHydOrder]{iterateHydOrder()}}}
}
\value{
nothing, function is called to generate graphical output
}
\description{
Present a \code{SoilProfileCollection} aligned to a geomorphic summary as cross-section.
}
\details{
Additional arguments to \code{\link[aqp:SoilProfileCollection-plotting-methods]{aqp::plotSPC()}} can be provided using \code{options(.aqp.plotSPC.args = list(...))}. For example, adjustments to maximum depth and profile width can be set via: \verb{options(.aqp.plotSPC.args = list(max.depth = 150, width = 0.35)}. Default arguments can be reset with \verb{options(.aqp.plotSPC.args = NULL}).

When \code{clust = TRUE}, especially for \code{SoilProfileCollections} with a wide range in depth, it may be necessary to adjust the \code{scaling.factor} argument to \code{\link[aqp:SoilProfileCollection-plotting-methods]{aqp::plotSPC()}} via: \code{options(.aqp.plotSPC.args = list(scaling.factor = 0.01))}. Larger values will increase the height of profile sketches.
}
\author{
D.E. Beaudette
}
