% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water-balance.R
\name{plotWB}
\alias{plotWB}
\title{Visualize Monthly Water Balance}
\usage{
plotWB(
  WB,
  AWC,
  fig.title = "",
  sw.col = "#377EB8",
  surplus.col = "#4DAF4A",
  et.col = "#E41A1C",
  deficit.col = "#FF7F00",
  pch = 22,
  pt.cex = 1,
  lwd = 1
)
}
\arguments{
\item{WB}{output from \code{monthlyWB}}

\item{AWC}{available water-holding capacity (mm)}

\item{fig.title}{a title}

\item{sw.col}{color for soil water}

\item{surplus.col}{color for surplus water}

\item{et.col}{color for ET}

\item{deficit.col}{color for deficit}

\item{pch}{plotting character for points (default: 22)}

\item{pt.cex}{character expansion factor for points (default: 1)}

\item{lwd}{line width (default: 1)}
}
\description{
This function offers one possible visualization for the results of \code{monthlyWB}.
}
\author{
D.E. Beaudette
}
\keyword{hplots}
