\name{PLSS2LL}
\alias{PLSS2LL}

\title{PLSS2LL}
\description{Fetch lattitude and longitude centroid coordinates for coded PLSS information from the BLM PLSS web service.}
\usage{PLSS2LL(p)}

\arguments{
  \item{p}{dataframe with chunks of PLSS coordinates}
}
\details{This function takes coded PLSS information and fetches the centroid lat/long coordinates down to the quarter-quarter section.  The 'plssid' column is generated within a dataframe using the following \code{\link{formatPLSS}} function.}
\value{A dataframe of PLSS codes and coordinates.}
\author{D.E. Beaudette, Jay Skovlin}
\note{This function expects that the dataframe will have a 'plssid' column generated by the \code{formatPLSS} function. Requires the following packages: \code{httr}, and \code{jsonlite}.}

\seealso{\code{\link{LL2PLSS}}, \code{\link{formatPLSS}}}

\keyword{manip}
