% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSS2LL.R
\name{PLSS2LL}
\alias{PLSS2LL}
\alias{PLSS2LL_1}
\alias{PLSS2LL_oneline}
\title{PLSS2LL}
\usage{
PLSS2LL(p)
}
\arguments{
\item{p}{data.frame with chunks of PLSS coordinates}

\item{plssid}{Column name containing PLSS ID (default: \code{"plssid"})}
}
\value{
A \code{data.frame} of PLSS codes and coordinates.
}
\description{
Fetch latitude and longitude centroid coordinates for coded PLSS information from the BLM PLSS web service.
}
\note{
This function expects that the dataframe will have a 'plssid' column generated by the \code{formatPLSS} function. Requires the following packages: \code{httr}, and \code{jsonlite}.
}
\examples{
if(requireNamespace("curl") &
   curl::has_internet()) {
  # create some data
  d <- data.frame(
    id = 1:3,
    qq = c('SW', 'SW', 'SE'),
    q = c('NE', 'NW', 'SE'),
    s = c(17, 32, 30),
    t = c('T36N', 'T35N', 'T35N'),
    r = c('R29W', 'R28W', 'R28W'),
    type = 'SN',
    m = 'MT20',
    stringsAsFactors = FALSE
  )
  
  # generate formatted PLSS codes
  d$plssid <- formatPLSS(d)
  
  # fetch lat/long coordinates
  PLSS2LL(d)
}
}
\seealso{
\code{\link{PLSS2LL_oneline}},\code{\link{LL2PLSS}}, \code{\link{formatPLSS}}
}
\author{
D.E. Beaudette, Jay Skovlin
}
