% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDailyClimateData.R
\name{prepareDailyClimateData}
\alias{prepareDailyClimateData}
\title{Prepare daily climate data (DAYMET)}
\usage{
prepareDailyClimateData(x, start, end, onlyWB = TRUE)
}
\arguments{
\item{x}{\code{SpatialPoints} object representing a single location}

\item{start}{start year (1998)}

\item{end}{end year (2018)}

\item{onlyWB}{logical, return just those date required by \code{dailyWB}}
}
\value{
a \code{data.frame}
}
\description{
Prepare daily climate data (DAYMET)
}
