% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilTaxonomyDendrogram.R
\name{SoilTaxonomyDendrogram}
\alias{SoilTaxonomyDendrogram}
\title{Soil Taxonomy Dendrogram}
\usage{
SoilTaxonomyDendrogram(
  spc,
  name = "hzname",
  name.style = "right-center",
  rotationOrder = NULL,
  max.depth = 150,
  n.depth.ticks = 6,
  scaling.factor = 0.015,
  cex.names = 0.75,
  cex.id = 0.75,
  axis.line.offset = -4,
  width = 0.1,
  y.offset = 0.5,
  shrink = FALSE,
  font.id = 2,
  cex.taxon.labels = 0.66,
  dend.color = par("fg"),
  dend.width = 1,
  ...
)
}
\arguments{
\item{spc}{a \code{SoilProfileCollection} object, typically returned by \code{soilDB::fetchOSD}}

\item{name}{column name containing horizon names}

\item{name.style}{passed to \code{aqp::plotSPC} (default: \code{"right-center"})}

\item{rotationOrder}{character vector of profile IDs with desired ordering of leaves in the dendrogram from left to right, ordering is not always possible}

\item{max.depth}{depth at which profiles are truncated for plotting}

\item{n.depth.ticks}{suggested number of ticks on the depth axis}

\item{scaling.factor}{scaling factor used to convert depth units into plotting units}

\item{cex.names}{character scaling for horizon names}

\item{cex.id}{character scaling for profile IDs}

\item{axis.line.offset}{horizontal offset for depth axis}

\item{width}{width of profiles}

\item{y.offset}{vertical offset between dendrogram and profiles}

\item{shrink}{logical, should long horizon names be shrunk by 80\% ?}

\item{font.id}{font style applied to profile id, default is 2 (bold)}

\item{cex.taxon.labels}{character scaling for taxonomic information}

\item{dend.color}{dendrogram line color}

\item{dend.width}{dendrogram line width}

\item{...}{additional arguments to \code{aqp::plotSPC}}
}
\value{
An invisibly-returned list containing:
\itemize{
\item \code{dist}: pair-wise dissimilarity matrix
\item \code{order}: final ordering of hclust leaves
}
}
\description{
Plot a dendrogram based on the first 4 levels of Soil Taxonomy, with soil profiles hanging below. A dissimilarity matrix is computed using Gower's distance metric for nominal-scale variables, based on order, sub order, great group, and subgroup level taxa. See the Details and Examples sections below for more information.
}
\details{
This function looks for specific site-level attributes named: \code{soilorder}, \code{suborder}, \code{greatgroup}, and \code{subgroup}. See \code{misc/soilTaxonomyDendrogram-examples.R} for some examples.

The \code{rotationOrder} argument uses (requires) the \code{dendextend::rotate()} function to re-order leaves within the \code{hclust} representation of the ST hierarchy. Perfect sorting is not always possible.
}
\examples{

# built-in data, same as results from soilDB::fetchOSD()
data("OSDexamples")

# use first 8 profiles
SoilTaxonomyDendrogram(
OSDexamples$SPC[1:8, ], width = 0.3, name.style = 'center-center'
)

}
\author{
D.E. Beaudette
}
