% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizTerracePosition.R
\name{vizTerracePosition}
\alias{vizTerracePosition}
\title{Visual Summary of Terraced Landform Positions}
\usage{
vizTerracePosition(x, s = NULL, annotations = TRUE, annotation.cex = 0.75)
}
\arguments{
\item{x}{\code{data.frame} as created by \code{soilDB::fetchOSD(..., extended=TRUE)}, see details}

\item{s}{an optional soil series name, highlighted in the figure}

\item{annotations}{logical, add number of record and normalized Shannon entropy values}

\item{annotation.cex}{annotation label scaling factor}
}
\value{
a \code{list} with the following elements:

\item{fig}{lattice object (the figure)}
\item{order}{ordering of soil series}
}
\description{
A unique display of terraced landform position probability.
}
\details{
See the \href{http://ncss-tech.github.io/AQP/soilDB/soil-series-query-functions.html}{Soil Series Query Functions} tutorial for more information.
}
\author{
D.E. Beaudette
}
